chore(serve): split out build logic
