+++ packages/serve/src/wrapper.marko
@@ -1,9 +1,9 @@
 $ const _flush = out.flush;
 $ out.flush = function () {
-    out.global.renderAssets(this);
+    input.renderAssets(this);
     _flush.apply(this, arguments);
     this.flush = _flush;
 }
 
 <${input.template}/>
-<${out.global.renderAssets}/>
+<${input.renderAssets}/>
